Scripts_ETL/requirements.txt
pandas==2.2.2
SQLAlchemy==2.0.29
psycopg2-binary==2.9.9
